Synthesist - Focus on Main Character
Long story short, I tried some weird things for the promo. However, I wasn't able to get it to work - As far as I can tell, the IsIncapacitated check fails for Hero controllers, since there are likely special multi-character pieces and the HTTC based check just fails.

So, it will be in the back of my mind, but in the meantime, all there is is the main version of the character.