Soulbinder Re-Implementation
After more thinking regarding mechanics and theme, I've started work on re-creating Soulbinder (originally Synthesist) as a full multi-character character. I believe the current mechanisms used should allow for the character to more gracefully handle things, by virtue of simply implementing everything as a multi-character to begin with. This was probably more complex than necessary, but I still want to see how well the planned mechanics work as is.

I've got the basic character loading down, so next will be the necessary implementation for the logic, as indicated by the setup.